refactor(crane-mcp): split skill-review.ts (815 LOC) into 4 modules (plan §L)
Extract three focused submodules under cli/skill-review/:
- frontmatter-parser.ts (145 LOC): YAML parser + Frontmatter type
- checks.ts (335 LOC): all 5 lint check functions + private helpers
- report.ts (57 LOC): types (Severity, Violation, ReviewResult) + formatters

Residual skill-review.ts (259 LOC): config loaders, reviewSkill orchestrator,
discoverSkillDirs, CLI arg parsing, main(), and barrel re-exports so
skill-review.test.ts imports are unchanged (zero test edits required).

All files <500 LOC. npm run verify passes. 25/25 skill-review tests pass.
